Move push commits from models to modules/repository
